read the description
I split the main notebook in "Analysis" and "Data Cleaning". the "data cleaning" notebook creates a new database with cleaned reviews, in string format. This means that in the analysis part they have to be tokenized again. For the cleaning I added a new function that checks for every token if it's a work or not, because now we still had things like "......." and other strange punctuation combinations in our reviews. The problem with that is that it takes a lot of time to run